Reworking the progress bar for each file to use a custom QLabel that does its own painting and implements the same API as the QProgressBar. Adding ability to directly edit the Mu value of the User Init areas in the QTableView. Eventually this will allow the user to directly manipulate the Gaussian curves to better fit the mixture of Gaussians model to the data.

+++ Code/EmMpmGui/EMMPMTask.cpp
@@ -70,8 +70,9 @@ void EMMPMTask::EMMPMUpdate_CallBackWrapper(EMMPM_Data* data)
   // warning: <globalSHPTask> MUST point to an appropriate object!
   EMMPMTask* mySelf = static_cast<EMMPMTask* > (data->userData);
 
+ // std::cout << "EMMPMUpdate_CallBackWrapper: Progress=" << data->progress << std::endl;
   emit
-  mySelf->progressValueChanged(data->progress);
+  mySelf->progressValueChanged((int)data->progress);
   emit
   mySelf->progressTextChanged(QString::number(data->progress));
